+++ packages/pliny/src/search/index.tsx
@@ -1,61 +1,34 @@
 import React from 'react'
-import dynamic from 'next/dynamic.js'
-// import { AlgoliaSearchContext } from './Algolia'
-// import { KBarContext } from 'kbar'
+import { AlgoliaSearchProvider } from './Algolia'
+import { KBarSearchProvider } from './KBar'
 
 import type { AlgoliaConfig } from './Algolia'
 import type { KBarConfig } from './KBar'
 
 export type SearchConfig = AlgoliaConfig | KBarConfig
-
 export interface SearchConfigProps {
   searchConfig: SearchConfig
   children: React.ReactNode
 }
 
-export interface SearchQuery {
-  setSearch: (search: string) => void
-  toggle: () => void
-}
-
-export interface SearchContext {
-  query: SearchQuery
-}
-
-const AlgoliaSearchProvider = dynamic(
-  () => {
-    return import('./Algolia').then((mod) => mod.AlgoliaSearchProvider)
-  },
-  { ssr: false }
-)
-
-const KBarSearchProvider = dynamic(
-  () => {
-    return import('./KBar').then((mod) => mod.KBarSearchProvider)
-  },
-  { ssr: false }
-)
-
-const KBarContext = dynamic(
-  // @ts-ignore
-  () => {
-    return import('kbar').then((mod) => mod.KBarContext)
-  },
-  { ssr: false }
-)
-
-const AlgoliaSearchContext = dynamic(
-  // @ts-ignore
-  () => {
-    return import('./Algolia').then((mod) => mod.AlgoliaSearchContext)
-  },
-  { ssr: false }
-)
-
 /**
  * Command palette like search component - `ctrl-k` to open the palette.
  * Or use the search context to bind toggle to an onOpen event.
  * Currently supports Algolia or Kbar (local search).
+ *
+ * To toggle the modal or search from child components, use the search context:
+ *
+ * For Algolia:
+ * ```
+ * import { AlgoliaSearchContext } from 'pliny/search/algolia'
+ * const { query } = useContext(AlgoliaSearchContext)
+ * ```
+ *
+ * For Kbar:
+ * ```
+ * import { useKBar } from 'kbar'
+ * const { query } = useKBar()
+ * ```
  *
  * @param {SearchConfig} searchConfig
  * @return {*}
@@ -79,16 +52,5 @@ export const SearchProvider = ({ searchConfig, children }: SearchConfigProps) =>
     }
   } else {
     return <>{children}</>
-  }
-}
-
-export const SearchContext = (provider: string): React.Context<SearchContext> => {
-  switch (provider) {
-    case 'algolia':
-      // @ts-ignore
-      return AlgoliaSearchContext
-    case 'kbar':
-      // @ts-ignore
-      return KBarContext
   }
 }